+++ js/data.js
@@ -1,0 +1,71 @@
+const numberPhoto = 25;
+
+const descriptions = [
+  'Солнечный закат',
+  'Цветущая природа',
+  'Горячий кофе',
+  'Снежный пейзаж',
+  'Волшебный лес',
+  'Туманный рассвет',
+  'Грозовая туча',
+  'Радужный дождь',
+  'Рыба-клоун аквариума',
+  'Красивый океан',
+  'Закрытый цветок',
+  'Мягкие облака',
+  'Луч солнца',
+  'Темный лес',
+  'Золотая осень',
+  'Белый снег',
+  'Бушующее море',
+  'Закатный огонь',
+  'Черный кошелек',
+  'Розовое облако',
+  'Лунная ночь',
+  'Радужный водопад',
+  'Солнечный цветок',
+  'Пышный ландшафт',
+  'Зеленый луг'
+];
+
+const names = [
+  'Анна',
+  'Борис',
+  'Владимир',
+  'Галина',
+  'Дмитрий',
+  'Евгений',
+  'Жанна',
+  'Зинаида',
+  'Ирина',
+  'Кирилл',
+  'Любовь',
+  'Мария',
+  'Никита',
+  'Ольга',
+  'Петр',
+  'Роман',
+  'Светлана',
+  'Татьяна',
+  'Ульяна',
+  'Федор',
+  'Христина',
+  'Цветана',
+  'Юлия',
+  'Яков',
+  'Яна'
+];
+
+const commentText = [
+  'Всё отлично!',
+  'В целом всё неплохо.',
+  'Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.',
+  'В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают.',
+  'Как можно было поймать такой неудачный момент?!'
+];
+
+export {numberPhoto, descriptions, names,commentText};
